Added a helper function the user model to get a user by id. Added my user sql dump.
Note About Auth->User()

We can never rely on the $this->Auth->User() information being all up to date. That data is only refreshed when the user authenticates because the Auth module just caches it. This is fine for the purpose of authentication because the that data (email, password) doesn't change. Or at least the user would have to reauthenticate. Data that does change like names, universities, twitter, fb ids. Are stored and changed in the db. If you change a value and try to access it through Auth->User() the data won't be up to date. So you should access that information which I like to call Profile information directly through the user model

So basically you should really use Auth->User() to get fields that you know won't change like the user.id or user.email.

In all other cases you can use the helper function I wrote which is $this->User->get($id) and it returns a user object
with all the associated fields as well (sublets and university information)