Ported over twitter verification functionality to the account page
refactored some of the code, copied verification code out of verifyController
and into accountController.

Exposed a bug in mine and tims database configuration. MySQl showtable bug
in our database.php files there is a switch statement at the bottom of it that did
some dev enviorment database config switching. It defaults to the dev database
which I had not defined. It was hitting the default case because after the twitter verification it'll
redirect to a url and supplying localhost to the oAuth redirect doesn't worked so you have to access
the url via 127.0.0.1 which localhost is an alias for. The case for accessing the app via this
address was not defined in the switch statement so it defaulted my setup to use an undefined setup.

I fixed it by making sure my database.php file handled that case, my solution was changing the default to use localhost