Facebook verification finally works!!!!!!!!!!!!!!!!!! HELL YEAH
Jesus that was a pain in the ass.
Javascript facebook verification works. Its not totally ajax yet
since there is a fair ammount of depencencies throughout the dashboard that references the user facebook id
for example the user profile icons which are added server side not client side.

But I ended up not using the Cookie method, the cookie method is more important for sharing a
state with your js sdk, a means for constantly having the users_id available. In our case we only want to
get the users id once and save it.

So I wrote an ajax function where I can post the 'signed request' received from the FB.login()
and then server side decrypt and verify the object. The user id can then be extracted and stored in the db