Resolved #15 removing listings
So when a sublet is deleted it is marked as invisible
the relevant marker will be made invisible as well if it no
longer has any visible properties referencing it.

When a sublet is added and Sublet->findMarkerById() if it finds
an existing marker that is invisible it will be visible again.

When loading the markers for the map I set a condition to only pull the visible markers

On the dashboard for the dropdown listings I only pull the visible ones now as well.