You can now have multiple date ranges for multiple listings.
Revamped the google wallet flow generation step. So the /getJwt action now supports
getting an array of listing_id, date_range pairs. These are now processed into a
new model called PendingOrder, this model basically holds a pending order while the user
completes the flow. On postback we will fufill the order. Before I was putting the order data
in the sellerdata field in the wallet jwt. This wasn't scalable as the payload it can hold
is only 200 characters.