Modified the order/postBackHandler to handle new order structure
So the postback handler works now, it takes in the jwt from a post request (I'm still faking it, real server soon)
and looks up the pending order corresponding to the jwt. We then go through all the items in the pending
order and fufill them. We then save a record of the order that includes a json encoded array of the items bought
in an order