Hopefully final refactor for the order flow.
Now have everything in place to support orders of any number and of any type
I refactored my coffeescript checkout module. I seperated out the featured listing order class,
which implements a orderItem interface that is used throughout the front end. Basically
it means that featured listing class supports the methods getOrderItem() and getPrice()
getOrderItem the object returned is a special object in that it is a generic item that can be bought.
It has the properties type, price, and item. The data under item is dependent on the type.
Its a shameful switch on type but I have found this hard to avoid as we are serializing data.

I moved the shopping cart logic and the buy logic into A2Cribs.Order, they are class functions.
They both take an array of orderItems and will either start the purchasing orderflow through wallet
or your items will be added to your cart.