cleaned up backend code, and fixed a shopping cart issue
went through and cleaned up a bit of the order back end code.
Removed duplicated logic and moved around some pieces.
Added more comments

Implemented a solution to the invalid cart problem.
Problem was what happens when the user has items in their shopping cart
and then one of them is invalid for example a featured listing order
has dates selected that are in the past or not in the valid range
The solution I came up with, naive as it is, is to make the carts
be reset every day.