changed the date format my datepicker was using to mysql format
less conversion needed now
also added the ability to add disabled dates to the datepicker
this can be used along with the getDates() function I added to featured
listings that will fetch all the dates a listing is featured. Combining
these two things we can, if we want, prevent users from double featuring
a listing as dates where the listing is aleady going to be featured will
be disabled.

I didn't add this functionality because there are a lot of other parts that need to change
such as the shopping cart since I need to do more lookups and I'm not sure how to not expose
all the dates a property is featured. Need to know if its safe to expose that.