Added in preventative front end validation for ordering featured listings
Wrote a new action in featured listing controller that returns a json
array of all the data needed to feature a listing. This includes
listing_id, street_address, alternate_name, listing_type (string),
as well as an array of unavailable dates that are specific to the listing
in question. I have all my backend validation in place but with this new
front end work, the only validation error that will be returned under
normal usage is the too many featured listings on one date error.
Other dates that could cause validation errors have been disabled on the
date picker so they'd have to manually add them and hack it in the console
to see if those issues