Changed FL_Dash to use the UserCache
Modified Dashboard.GetListings to return a deffered to interested parties
can know when the Cache gets fetched (my FL_Dash).

Also modified my FeaturedListings GetUnavailableDates function to return
an array of full dates as well as a by listing_id map of arrays of dates
that the listing is already featured on.