Finally got order items being correctly generate in the fl dash
I've been reworking a lot the fl dash order logic to accomedate the a la carte
university selection. I've got it working now so when the user hits buy
it'll generate order items that contain listing_id, uni_id, and dates
the dates array contains the dates selected without the unavailable dates
so if for example. The user features a listing at UM on one day.
They come back and select a range that includes that day and want to feature
it at 3 different places including UM, they will be told its already featured
at UM and the price will reflect it not being featured again on that date
when the hit buy that date will also not be included.

I initially thought about doing this removal on the backend but that would slop up
my order flow. Now on the back end I receive an array of order items. That contain
a listing_id, uni_id and dates. If all the flow went correctly no date should be
in that array that its already featured at. If I did this date removal business on the
back end it would complicate my validation code because we wanted to warn the user
but not invalidate the order. Now since its done on the front end the user
will know immediately if the listing has a conflict but it won't impeded their
ordering process